Move and rename fibaro base entity to separate module (#126055)
