+++ homeassistant/components/fibaro/lock.py
@@ -12,8 +12,9 @@ from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import FibaroController, FibaroDevice
+from . import FibaroController
 from .const import DOMAIN
+from .entity import FibaroEntity
 
 
 async def async_setup_entry(
@@ -29,7 +30,7 @@ async def async_setup_entry(
     )
 
 
-class FibaroLock(FibaroDevice, LockEntity):
+class FibaroLock(FibaroEntity, LockEntity):
     """Representation of a Fibaro Lock."""
 
     def __init__(self, fibaro_device: DeviceModel) -> None:
